python/src/auth/requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
